Dockerfile: remove Zephyr SDK installation
The build system provides a prebuilt toolchain to use instead, along
with some other changes to the genesis-dev metapackage which pull in
other dependencies provided by the Zephyr SDK.

Since that implies the Zephyr SDK is no longer needed, don't install
it. This removes a layer and decreases both the build and download
size of the container.

Signed-off-by: Marti Bolivar <marti.bolivar@linaro.org>