Dockerfile: clean up python installation dependencies
The pip3 that comes with ubuntu:16.04 isn't able to deal with wheels.
This is now causing a build error when installing the cryptography
dependency.

Fix it by upgrading pip, installing wheel, and then installing our
dependencies. This allows us to install the wheel for pyelftools as
well.

We need to drop the --system argument, which was removed from later
versions of pip.  This is fine; --user is the default on Debian
systems only when run as non-root.

Signed-off-by: Marti Bolivar <marti@opensourcefoundries.com>