Dockerfile: add workaround for pip v10 related breakage
After upgrading pip, we need to clear that Bash command cache in order
to ensure the upgraded version is what is discovered on PATH when we
run pip3 in subsequent lines. Otherwise, an import error is raised
when we run the system pip3, which uses internal pip APIs that changed
in v10.

Signed-off-by: Marti Bolivar <marti@opensourcefoundries.com>